#308 split reading comments to DfpropSchemaPolicyFileCommentLogic
